metodos, act. modelo, informe... falta arreglar clase informe, mañana sigo

+++ BibliotecaControlador/AGENDA.cs
@@ -19,10 +19,10 @@ namespace BibliotecaDALC
             this.SOLICITUD = new HashSet<SOLICITUD>();
         }
     
-        public decimal ID_AGENDA { get; set; }
+        public int ID_AGENDA { get; set; }
         public System.DateTime DIA { get; set; }
         public string HORA { get; set; }
-        public decimal ID_EQUIPO { get; set; }
+        public int ID_EQUIPO { get; set; }
     
         public virtual EQUIPO_TECNICO EQUIPO_TECNICO { get; set; }
         public virtual ICollection<SOLICITUD> SOLICITUD { get; set; }
