+++ BibliotecaControlador/TECNICO.cs
@@ -25,10 +25,10 @@ namespace BibliotecaDALC
         public string AP_PATERNO { get; set; }
         public string AP_MATERNO { get; set; }
         public string DIRECCION { get; set; }
-        public decimal TELEFONO { get; set; }
+        public int TELEFONO { get; set; }
         public string EMAIL { get; set; }
-        public decimal ID_EQUIPO { get; set; }
-        public decimal ID_COMUNA { get; set; }
+        public int ID_EQUIPO { get; set; }
+        public int ID_COMUNA { get; set; }
     
         public virtual COMUNA COMUNA { get; set; }
         public virtual EQUIPO_TECNICO EQUIPO_TECNICO { get; set; }
